+++ builder/ui/buildListItem.ts
@@ -1,5 +1,0 @@
-export const buildListItem = (content, type: string) => {
-  if (type === "md") return `- ${content}`;
-  if (type === "html") return `<li>${content}</li>`;
-  return "";
-};