+++ builder3Server/builder/pandoc.ts
@@ -12,7 +12,7 @@ export class Pandoc {
   constructor() {
   }
 
-  async generate(input: PandocInput) {
+  async generate(input: PandocInput): Promise<void> {
     const cmd = this.createCommand(input);
     const { stdout, stderr } = await exec(cmd);
     console.log("stdout:", stdout);