+++ builder3Server/server/utils/fetchBodyParams.ts
@@ -1,0 +1,5 @@
+import { Request } from "express";
+
+export const fetchBodyParams = (req: Request): any => {
+  return req.body;
+};