+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 * add prettier and eslint settings
 * add installation documentation
 * add support of pandoc metadata
+* add wkhtmltopdf support
 
 # [0.0.74] - 2024-06-10
 * add controllers concept to server