+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 * add wkhtmltopdf support
 * fix bug of support latex formulas
 * fix bug of a space between title and a table of contents
+* remove Builder2
 
 # [0.0.74] - 2024-06-10
 * add controllers concept to server