+++ changelog.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 * fix bug of support latex formulas
 * fix bug of a space between title and a table of contents
 * remove Builder2
+* remove legacy tooling and metadata
 
 # [0.0.74] - 2024-06-10
 * add controllers concept to server