+++ changelog.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * fix bug of a space between title and a table of contents
 * remove Builder2
 * remove legacy tooling and metadata
+* add Understanding Inversion of Control (IoC) article
 
 # [0.0.74] - 2024-06-10
 * add controllers concept to server